+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # v2.23.5
 
+* added AutoflushFile.seek()
+* added AutoflushFile.truncate()